// src/utils/interfaces.ts
export interface SpaceValidation {
  name: string;
  params: Record<string, any>;
}

export interface SpaceStrategy {
  name: string;
  network: string;
  params: Record<string, unknown>;
}
// ...
export interface VoteValidation {
  name: string;
  params: Record<string, any>;
}